COMPLETED Logout feature

+++ src/main/java/io/github/anantharajuc/rc/authentication/service/RefreshTokenServiceImpl.java
@@ -1,5 +1,6 @@
 package io.github.anantharajuc.rc.authentication.service;
 
+import java.util.Optional;
 import java.util.UUID;
 
 import org.springframework.stereotype.Service;
@@ -9,20 +10,23 @@ import io.github.anantharajuc.rc.authentication.model.RefreshToken;
 import io.github.anantharajuc.rc.authentication.repository.RefreshTokenRepository;
 import io.github.anantharajuc.rc.exceptions.SpringRedditException;
 import lombok.AllArgsConstructor;
+import lombok.extern.log4j.Log4j2;
 
 @Service
 @AllArgsConstructor
 @Transactional
+@Log4j2
 public class RefreshTokenServiceImpl implements RefreshTokenService
 {
 	private final RefreshTokenRepository refreshTokenRepository;
 	
 	@Override
-	public RefreshToken generateRefreshToken() 
+	public RefreshToken generateRefreshToken(String username) 
 	{
 		 RefreshToken refreshToken = new RefreshToken(); 
 		 
 		 refreshToken.setToken(UUID.randomUUID().toString());
+		 refreshToken.setUsername(username); 
 		 
 		 return refreshTokenRepository.save(refreshToken);
 	}
@@ -34,8 +38,29 @@ public class RefreshTokenServiceImpl implements RefreshTokenService
 	}
 
 	@Override
-	public void deleteByToken(String token) 
+	public void deleteByToken(String token, String username) 
 	{
-		refreshTokenRepository.deleteByToken(token); 
+		log.info("Delete by token method start");
+		
+		RefreshToken refreshToken = refreshTokenRepository.findByToken(token).orElseThrow(() -> new SpringRedditException("Invalid Refresh Token"));
+		//Optional<RefreshToken> refreshToken = refreshTokenRepository.findByToken(token);
+		
+		//log.info(refreshToken.isPresent());
+		//log.info(refreshToken.get().getToken().equals(token));
+		//log.info(refreshToken.get().getUsername().equals(username));
+		
+		
+		if(refreshToken.getToken().equals(token) && refreshToken.getUsername().equals(username)) 
+		{
+			log.info("Delete by token condition satisfied");	
+				
+			refreshTokenRepository.deleteByToken(token); 
+		}
+		else
+		{
+			log.info("Token, Username mismatch!");	
+			
+			throw new SpringRedditException("Token, Username mismatch!");
+		}
 	}
 }
